requirements.txt
openai==1.57.4
praw==7.8.1
python-dotenv==1.0.1
